refactor code to use const instead of var

+++ app/Controllers/Http/DeleteTController.ts
@@ -4,14 +4,13 @@ import Tarantula from 'App/Models/Tarantula'
 export default class DeleteTController {
   public async post(ctx: HttpContextContract) {
     try {
-      var tarantula = await Tarantula.findOrFail(ctx.params.id)
-      if (!ctx.auth.user || tarantula.user_id != ctx.auth.user.id) {
-        throw new Error('Unauthorized')
+      const tarantula = await Tarantula.find(ctx.params.id)
+      if (!tarantula) {
+        return ctx.response.badRequest('Bad Request')
       }
-    } catch (error) {
-      return ctx.response.badRequest('Bad request')
-    }
-    try {
+      if (!ctx.auth.user || tarantula.user_id != ctx.auth.user.id) {
+        return ctx.response.unauthorized('Unauthorized')
+      }
       await tarantula.delete()
       ctx.session.flash('global_message', 'Tarantula deleted')
       return ctx.response.redirect('/user/home')
